dab/requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aiosignal==1.3.1
attrs==24.2.0
blinker==1.8.2
click==8.1.7
discord.py==2.4.0
Flask==3.0.3
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.1
Werkzeug==3.0.3
yarl==1.9.4
